feat: all of the overdue date will display a red message

+++ src/main.ts
@@ -11,6 +11,8 @@ const due_date = document.querySelector<HTMLInputElement>('#due-date')
 const error_message = document.querySelector<HTMLParagraphElement>(
   '#todo-creation-error',
 )
+const global_message =
+  document.querySelector<HTMLParagraphElement>('#global_message')
 
 interface Todo {
   text: string
@@ -33,6 +35,7 @@ function done_todo(index: number) {
     }
   }
   localStorage.setItem('value', JSON.stringify(todos))
+  overdueTodos()
 }
 
 function button_disabled() {
@@ -44,6 +47,18 @@ function button_disabled() {
     }
 }
 
+function overdueTodos() {
+  const today = new Date()
+  const overdueTodos = todos.filter((todo) => new Date(todo.date) < today)
+
+  if (overdueTodos.length > 0 && global_message) {
+    global_message.innerHTML =
+      "<p style='background-color: red; color: white; '>You have overdue todos !</p>"
+  } else if (global_message) {
+    global_message.innerHTML = ''
+  }
+}
+
 function myList(todo: Todo, index: number) {
   if (outputList) {
     const addedTodoText = todo.text
@@ -51,12 +66,13 @@ function myList(todo: Todo, index: number) {
     newList.textContent = addedTodoText
     outputList.appendChild(newList)
 
-    if (delete_all) {
+    if (delete_all && global_message) {
       delete_all.addEventListener('click', () => {
         localStorage.removeItem('value')
         newList.remove()
         button.remove()
         checkbox.remove()
+        global_message.innerHTML = ''
       })
     }
     const TextRemove = 'Remove'
@@ -83,6 +99,7 @@ function myList(todo: Todo, index: number) {
       if (index !== -1) {
         todos.splice(index, 1)
         localStorage.setItem('value', JSON.stringify(todos))
+        overdueTodos()
       }
     })
 
@@ -109,15 +126,15 @@ function myList(todo: Todo, index: number) {
           new Date(todo.date).setHours(0, 0, 0, 0) ===
           new Date().setHours(0, 0, 0, 0)
         ) {
-          time.style.color = 'orange'
+          newList.style.color = 'orange'
         } else if (new Date(todo.date) < new Date()) {
-          time.style.color = 'red'
+          newList.style.color = 'red'
         } else if (
           new Date(todo.date) < new Date(today.setDate(today.getDate() + 4))
         ) {
-          time.style.color = 'yellow'
+          newList.style.color = 'yellow'
         } else {
-          time.style.color = 'green'
+          newList.style.color = 'green'
         }
 
         dates.appendChild(time)
@@ -128,8 +145,6 @@ function myList(todo: Todo, index: number) {
   } else {
     alert('Please enter a todo ')
   }
-
-  // call setHours to take the time out of the comparison
 }
 
 if (todoInput) {
@@ -148,7 +163,6 @@ if (button && todoInput && outputList) {
   })
 }
 
-// Stock la todo
 function test(): void {
   if (todoInput && due_date) {
     const text: string = todoInput.value.trim()
@@ -162,4 +176,6 @@ function test(): void {
       todoInput.value = ' '
     }
   }
+  overdueTodos()
 }
+overdueTodos()
